+++ GroceryStore-Backend/src/main/java/ca/mcgill/ecse321/GroceryStoreBackend/dao/ReviewRepository.java
@@ -12,6 +12,6 @@ public interface ReviewRepository extends CrudRepository<Review, Integer> {
 
   Review findReviewByCustomerAndOrder(Customer customerEmail, Order orderId);
   
-  Review findReviewById(Integer reviewId);
+  Review findReviewById(Long reviewId);
 
 }
